TINY-11394: revert `SizeInput` to the version pre `7.6`
